requirements.txt
python-telegram-bot==20.7
fastapi==0.110.0
uvicorn==0.29.0
yookassa==2.3.2